Add Migration

+++ Xurmo.Domain/Entities/OrderItem.cs
@@ -6,8 +6,8 @@ public class OrderItem : Auditable
 {
     public int Quantity { get; set; }
     public decimal Amount { get; set; }
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
     public Order Order { get; set; }
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     public Product Product { get; set; }
 }
